refactor(imagetypes): replace NamedTuple with dataclass for ImageGeometry and add Direction class
