+++ src/imgtools/coretypes/helper_types.py
@@ -11,38 +11,29 @@ spatial transformations, bounding box calculations, and metadata representation.
   and reflect their real-world meaning.
 - **Reusability:** Types should be generic enough to apply to various spatial
   operations across domains, especially medical imaging.
-- **Consistency:** Ensure compatibility with spatial metadata in imaging formats
-  (e.g., DICOM, NIfTI).
-- **Extendability:** Enable seamless addition of new spatial concepts as needed.
 
 ## Types:
-1. **Point3D**
-   - Represents a point in 3D space with x, y, z coordinates.
+1. **Vector3D**
+   - Represents a vector in 3D space with x, y, z components.
    - Includes methods for basic vector arithmetic (addition, subtraction).
 
 2. **Size3D**
    - Represents the dimensions of a 3D object (width, height, depth).
-   - Derived from `Point3D` but semantically indicates size rather than position.
+   - Includes methods to calculate volume.
 
-3. **Coordinate**
+3. **Coordinate3D**
    - Represents a specific coordinate in 3D space.
-   - Identical in structure to `Point3D` but used to denote spatial locations.
+   - Inherits from `Vector3D` and includes methods for addition and subtraction
+     with other `Coordinate3D` or `Size3D` objects.
 
-4. **Direction**
-   - Represents a directional vector in 3D space, useful for orientation.
-   - Includes normalization and utility functions to ensure consistency with
-     medical imaging metadata like direction cosines.
+4. **Spacing3D**
+   - Represents the spacing in 3D space.
+   - Inherits from `Vector3D`.
 
-These types serve as building blocks for higher-level spatial operations and
-enhance code readability, ensuring that spatial concepts are represented
-accurately and intuitively.
-
-Examples:
----------
->>> point = Point3D(x=10, y=20, z=30)
->>> size = Size3D(x=50, y=60, z=70)
->>> coord = Coordinate(x=5, y=5, z=5)
->>> direction =
+5. **Direction**
+   - Represents a directional matrix for image orientation.
+   - Includes methods for normalization and conversion between flattened and
+     matrix forms.
 """
 
 from __future__ import annotations
@@ -54,27 +45,6 @@ from dataclasses import dataclass
 from typing import NamedTuple, Sequence
 from typing import Tuple, Iterator, TypeAlias
 import numpy as np
-
-Matrix3D: TypeAlias = Tuple[
-    Tuple[float, float, float],
-    Tuple[float, float, float],
-    Tuple[float, float, float],
-]
-Matrix3DFlat: TypeAlias = Tuple[
-    float, float, float, float, float, float, float, float, float
-]
-
-FlattenedMatrix = Matrix3DFlat
-
-
-# @dataclass(frozen=True)
-# class ImageGeometry:
-#     """Represents the geometry of a 3D image."""
-
-#     size: Size3D
-#     origin: Point3D
-#     direction: Direction
-#     spacing: Spacing
 
 
 @dataclass
@@ -99,6 +69,8 @@ class Vector3D:
         Subtract another Vector3D from this vector.
     __iter__():
         Iterate over the components (x, y, z).
+    __getitem__(index):
+        Access components via index.
     """
 
     x: float
@@ -123,13 +95,25 @@ class Vector3D:
         """Allow iteration over the components."""
         return iter((self.x, self.y, self.z))
 
+    def __getitem__(self, idx: int) -> float:
+        """Access components via index."""
+        # return (self.x, self.y, self.z)[index]
+        match idx:
+            case int() as index:
+                return (self.x, self.y, self.z)[idx]
+            case str() if idx in vars(self):
+                return getattr(self, idx)
+            case _:
+                errmsg = f"Invalid index: {idx}."
+                raise IndexError(errmsg)
+
     def __repr__(self) -> str:
         """Return a string representation of the Vector3D."""
         cls = self.__class__.__name__
         return f"{cls}(x={self.x:.1f}, y={self.y:.1f}, z={self.z:.1f})"
 
 
-class Spacing(Vector3D):
+class Spacing3D(Vector3D):
     """
     Represent the spacing in 3D space.
     Inherits from Vector3D.
@@ -240,112 +224,6 @@ class Size3D:
         )
 
 
-"""
-jermiah (01/30/2025): 
-not sure about this direction class, was an initial proof-of-concept
-doesnt seem like a dealbreaker to need, havent dealt with any issues
-Really only wanted to use it for docs/tutorials to show how images
-are oriented + work via directions for sitk.Image
-
-lets not consider it too much unless we need to debug with it
-"""
-
-
-@dataclass(frozen=True, eq=True)
-class Direction:
-    """
-    Represent a directional matrix for image orientation.
-
-    Supports 3D (3x3) directional matrices in row-major format.
-
-    Attributes
-    ----------
-    matrix : Tuple[float, ...]
-        A flattened 1D array representing the matrix,
-        with length 9 (3x3).
-    """
-
-    matrix: Matrix3DFlat
-
-    def __post_init__(self) -> None:
-        length = len(self.matrix)
-        if length != 9:
-            msg = (
-                "Direction must be a 3x3 (9 values) matrix."
-                f" Got {length} values."
-            )
-            raise ValueError(msg)
-
-    @classmethod
-    def from_matrix(
-        cls,
-        matrix: Matrix3D,
-    ) -> Direction:
-        """
-        Create a Direction object from a full 3D (3x3) matrix.
-
-        Parameters
-        ----------
-        matrix : Matrix3D
-            A nested tuple representing the 3D matrix.
-
-        Returns
-        -------
-        Direction
-            A Direction instance.
-        """
-        if (size := len(matrix)) != 3:
-            msg = f"Matrix must be 3x3. Got {size=}."
-            raise ValueError(msg)
-        for row in matrix:
-            if len(row) != size:
-                raise ValueError("Matrix must be square (3x3).")
-        flattened: FlattenedMatrix = tuple(
-            value for row in matrix for value in row
-        )  # type: ignore
-        return cls(matrix=flattened)
-
-    def to_matrix(self) -> list[list[float]]:
-        """Convert the flattened row-major array back to a 3D matrix."""
-        dim = 3
-        return [list(self.matrix[i * dim : (i + 1) * dim]) for i in range(dim)]
-
-    def normalize(self) -> Direction:
-        """Return a new Direction with normalized row vectors."""
-        matrix = self.to_matrix()
-        normalized_matrix = [
-            list(np.array(row) / np.linalg.norm(row)) for row in matrix
-        ]
-        return Direction.from_matrix(
-            tuple(tuple(row) for row in normalized_matrix)  # type: ignore
-        )
-
-    def is_normalized(self, tol: float = 1e-6) -> bool:
-        """Check if the row vectors of the matrix are normalized."""
-        matrix = self.to_matrix()
-        for row in matrix:
-            if not np.isclose(np.linalg.norm(row), 1.0, atol=tol):
-                return False
-        return True
-
-    def __iter__(self) -> Iterator:
-        """Allow the Direction instance to be passed directly as a 1D array."""
-        yield from self.matrix
-
-    def __repr__(self) -> str:
-        dim = 3
-        rows = self.to_matrix()
-        formatted_rows = [
-            "  [" + ", ".join(f"{value:>7.3f}" for value in row) + "]"
-            for row in rows
-        ]
-        return (
-            f"Direction(  {dim}x{dim} matrix\n"
-            + "\n".join(formatted_rows)
-            + f"\n)"
-        )
-
-
 if __name__ == "__main__":
     # ruff : noqa
     from pathlib import Path
@@ -357,13 +235,18 @@ if __name__ == "__main__":
 
     vector1 = Vector3D(1.0, 2.0, 3.0)
 
-    vector2 = Vector3D((1.0, 2.0, 3.0))  # as tuple input
+    vector2 = Vector3D((1, 2, 3))  # as tuple input
     assert all((attr1 == attr2) for attr1, attr2 in zip(vector1, vector2))
 
+    # iterate over the objects' attributes
     for attr in vector1:
         print(attr)
 
     print(vector1)
+
+    # access object attributes via [ ]
+    print(f"{vector1[0]}")
+    print(f"{vector1['y']}")
 
     ########################################
     # Coordinate3D
@@ -381,52 +264,3 @@ if __name__ == "__main__":
 
     print(f"Adding {point=} and {size_tuple=} = {point_plus_size}")
     print(f"Adding {point=} and {size=} = {point + size}")
-
-    # Create instances of each class
-    # point = Point3D(10.0, 20.0, 30.0)
-    # size = Size3D(50.0, 60.0, 70.0)
-    # direction = Direction.from_matrix(
-    #     (
-    #         (0.707, 0.707, 0.0),
-    #         (-0.707, 0.707, 0.0),
-    #         (0.0, 0.0, 1.0),
-    #     )
-    # )
-
-    # # Testing Point3D and Size3D operations
-    # new_point = point + size
-
-    # # Printing out the details
-    # print(f"Point: {point}")
-    # print(f"Size: {size}")
-    # print(f"New point after adding size: {new_point}")
-    # print(f"Direction: {direction}")
-
-    # # Unpacking the values
-    # print("Unpacked Point:", tuple(point))  # (10.0, 20.0, 30.0)
-    # print("Unpacked Size:", tuple(size))  # (50.0, 60.0, 70.0)
-
-    # example_image = np.random.rand(10, 10, 10)
-    # example_sitk_image = sitk.GetImageFromArray(example_image)
-
-    # # Create a direction vector
-    # # 3x3 Direction Matrix
-    # direction_3d = Direction.from_matrix(
-    #     (
-    #         (0.707, 0.707, 0.0),
-    #         (-0.707, 0.707, 0.0),
-    #         (0.0, 0.0, 1.0),
-    #     )
-    # )
-
-    # print(f"{direction_3d=}")
-    # example_sitk_image.SetDirection(direction_3d)
-
-    # # make another direction from a flattened
-    # direction_3d_flat = Direction(
-    #     matrix=(0.707, 0.707, 0.0, -0.707, 0.707, 0.0, 0.0, 0.0, 1.0)
-    # )
-    # print(f"{direction_3d_flat=}")
-    # print(f"{(direction_3d_flat==direction_3d)=}")
-
-    # print(f"{example_sitk_image=}")
